pyproject.toml
[tool.poetry]
name = "grpc-helloworld"
version = "0.1.0"
description = ""
authors = ["daggy1234"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.56.2"
grpcio-tools = "^1.56.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
